requirements.txt
asgiref==3.6.0
autopep8==2.0.1
Django==4.1.5
django-filter==22.1
Pillow==9.4.0
psycopg2==2.9.5
pycodestyle==2.10.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7